requirements.txt
pymupdf>=1.24.0
pillow>=10.0.0
opencv-python>=4.8.0
faker>=20.0.0
rich>=13.0.0
numpy>=1.24.0
